DevopsClass/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
django-cors-headers
dj-database-url